Address feedback based on other PR
